Editor: Improve frontend performance for `get_default_block_editor_settings()`.
The `wp_max_upload_size()` function can be expensive to call, especially for large sites or multisites. For the frontend usage of `get_default_block_editor_settings()` knowing the allowed upload size is typically unnecessary.

This changeset adds a condition so that `wp_max_upload_size()` is only called if the current user can actually `upload_files`. It keeps the data present when it is actually needed while avoiding the execution overhead when it is not needed.

Props janthiel, Clorith, flixos90, spacedmonkey.
Merges [54769] to the 6.1 branch.
Fixes #56815.

Built from https://develop.svn.wordpress.org/branches/6.1@54772


git-svn-id: https://core.svn.wordpress.org/branches/6.1@54324 1a063a9b-81f0-0310-95a4-ce76da25c4cd