Upgrade/Install: Delete the `comments-query-loop` folder in `wp-includes/blocks`.
This properly deletes the now empty `src/wp-includes/blocks/comments-query-loop` directory and adds that directory to the `$_old_files` array.

The files in this directory were removed in [54257], but the directory was not marked as deleted in SVN.

Props azaozz, jorbin, SergeyBiryukov.
Merges [54836] to the 6.1 branch.
Fixes #57080.
Built from https://develop.svn.wordpress.org/branches/6.1@54837


git-svn-id: https://core.svn.wordpress.org/branches/6.1@54389 1a063a9b-81f0-0310-95a4-ce76da25c4cd