Help/About: Update “Get involved in WordPress” link on the Credits tab.
This commit updates the link URL to send contributors to an orientation tool for a short survey on what team they may be interested in joining. This coincides with the Get Involved tab.

Follow-up to [17877], [26354], [35898], [43032], [56220].

Props courane01, audrasjb.
Reviewed by audrasjb, SergeyBiryukov.
Merges [56311] to the 6.3 branch.
See #23348.
Built from https://develop.svn.wordpress.org/branches/6.3@56312


git-svn-id: https://core.svn.wordpress.org/branches/6.3@55824 1a063a9b-81f0-0310-95a4-ce76da25c4cd