General: Commit changes after running `precommit:css`.
The `{min|max}-device-pixel-ratio` syntax as a a non-standard CSS media feature that was used as an alternative to the now standard `resolution`. Prior to Safari 16.0, `-webkit-{min|max}-device-pixel-ratio` was needed to correctly support it.

This change is a result of the `caniuse-lite` update that was applied recently in [56065]. Though there were no changes to target browsers as a result of this update, it seems an upstream change identified these as unnecessary.

Props desrosj, joemcgill, isabel_brison, SergeyBiryukov, azaozz.
Reviewed by SergeyBiryukov, azaozz.
Backports [56279] to the 6.3 branch.
Fixes #57856.

Built from https://develop.svn.wordpress.org/branches/6.3@56367


git-svn-id: https://core.svn.wordpress.org/branches/6.3@55879 1a063a9b-81f0-0310-95a4-ce76da25c4cd