Build/Test Tools: Change the default value of `LOCAL_PHP`.
This changes the default value for `LOCAL_PHP` in the 6.4 branch from `latest` to `8.3-fpm` to reflect the highest version of PHP this branch will support (with beta support).

After this change, future updates to the `latest` container built and published by the `wpdev-docker-images` repository will not cause failures in this branch.

While there are currently no failures in the 6.4 branch, the `latest` container was recently changed from using PHP 7.4 to 8.2 and caused failures in the 5.8-6.3 branches. Pinning a PHP version to this branch now will prevent that same scenario in the future. Going forward, this should be done at the time of branching.

Older branches already have specific PHP versions defined.

See #60095.
Built from https://develop.svn.wordpress.org/branches/6.4@57198


git-svn-id: https://core.svn.wordpress.org/branches/6.4@56709 1a063a9b-81f0-0310-95a4-ce76da25c4cd