Tests: Update expectations in `wp_remote_head()` and `wp_remote_get()` tests.
It appears that something has changed on the WP.com side to compress the requested images on the fly, which interfered with the previous expectations in these tests.

Follow-up to [139/tests], [31258], [47142].

Reviewed by audrasjb.
Merges [57903] and [57904] to the 6.5 branch.
Props dextorlobo, swissspidy, davidbaumwald, SergeyBiryukov.
Fixes #60865.



Built from https://develop.svn.wordpress.org/branches/6.5@57908


git-svn-id: https://core.svn.wordpress.org/branches/6.5@57409 1a063a9b-81f0-0310-95a4-ce76da25c4cd