Twenty-Twenty: Fixes space between post content on front.
The post author and post date did not have space between them and the post content. This brings in 1em of top margin. Of note is that this only is if the first element is a paragraph that the issue was caused.  
  
Props abcd95, sabernhardt, desrosj, sainathpoojary, viralsampat.  
Fixes #62243.

Built from https://develop.svn.wordpress.org/trunk@59470


git-svn-id: https://core.svn.wordpress.org/trunk@58856 1a063a9b-81f0-0310-95a4-ce76da25c4cd