Build/Test Tools: Allow more control when testing older branches.
This adds an input to the Test Old Branches workflow that allows a specific branch to be specified or `all` to run all old branches.

The default behavior is to only test the currently supported version of WordPress as defined in the `CURRENTLY_SUPPORTED_BRANCH` environment variable.

Follow up to [59520].

See #62221.
Built from https://develop.svn.wordpress.org/trunk@59529


git-svn-id: https://core.svn.wordpress.org/trunk@58915 1a063a9b-81f0-0310-95a4-ce76da25c4cd