Build/Test Tools: Switch to using local references for reusable workflows.
The benefit of this is that when PRs are made to make changes to a reusable workflow, the references doesn't need to be updated to point to the fork in order for the changed workflow to run.

A `npm run grunt replace:workflow-references-local-to-remote` command has also been introduced in order to convert these local references back to remote ones. This command can be used to switch release branches over to using remote workflows, as they are currently, so they continue to benefit from workflow changes in trunk without the need for continual backporting to all the branches.

Props desrosj, johnbillion

Fixes #62416

Built from https://develop.svn.wordpress.org/trunk@59673


git-svn-id: https://core.svn.wordpress.org/trunk@59016 1a063a9b-81f0-0310-95a4-ce76da25c4cd