Build/Test Tools: Correct the usage of the conclusion of the previous run within the Slack notifications workflow.
Prior to r59679 this value was echoed to GitHub output which meant that its surrounding double quotes lost their significance. Now this value is used directly in the job output it needs to be treated as a plain string.

This concludes the conclusion confusion.

See #82221

Built from https://develop.svn.wordpress.org/trunk@59693


git-svn-id: https://core.svn.wordpress.org/trunk@59035 1a063a9b-81f0-0310-95a4-ce76da25c4cd