Build/Test Tools: Avoid using `*-latest` tags for runner images.
While using the `ubuntu-latest`, `macos-latest`, and `windows-latest` runner image tags is convenient, it has proven to be problematic in a number of instances as the runners are slowly updated (see #62808 and #62843).

This switches all workflows to using specific version tags representing the latest non-preview versions, which currently are as follows:
- `ubuntu-24.04`
- `windows-2022`
- `macos-14`

Props swissspidy, johnbillion.
See #62221.
Built from https://develop.svn.wordpress.org/trunk@59720


git-svn-id: https://core.svn.wordpress.org/trunk@59062 1a063a9b-81f0-0310-95a4-ce76da25c4cd