External Libraries: Append .1 to react/react-dom versions.
In 6.7, [58775] changed the way react and react-dom are bundled in Core. This commit resulted in some changes to the built files that are distributed in WordPress even though the actual version of the libraries remained the same. The result can be a blank white screen when trying to edit a post when those two script files are heavily cached.

This uses `.1` as the appendage to prevent issues with third party code expecting the version number in the form `/^[\d\.]+$/`.

Updates include tests in `Tests_Dependencies_Scripts::test_vendor_script_versions_registered_manually` to include the modified version used for cache busting.

Reviewed by jorbin.
Merges [59536] and [59540] to 6.7 branch.

Props levskipg, get_dave, smerriman, jdnd, juanwp22, seanlanglands, robertstaddon, azaozz, peterwilsoncc, desrosj.
Fixes #62422.

Built from https://develop.svn.wordpress.org/branches/6.7@59741


git-svn-id: https://core.svn.wordpress.org/branches/6.7@59083 1a063a9b-81f0-0310-95a4-ce76da25c4cd