Build/Test Tools: Prevent Dependabot workflow runs in other mirrors.
Currently, workflows are configured to only run for `wordpress-develop` or when pull requests are opened where forks and mirrors are the base repository.

Because a Dependabot configuration is present in the repository, it cannot be turned off for mirrors. This results in workflows running for all Dependabot PRs, which is problematic for private mirrors and needlessly consumes minutes and resources.

Props swissspidy, johnbillion.
See #62221.
Built from https://develop.svn.wordpress.org/trunk@60080


git-svn-id: https://core.svn.wordpress.org/trunk@59416 1a063a9b-81f0-0310-95a4-ce76da25c4cd