General: Adjust `WP_Community_Events::format_event_data_time()` deprecation.
`_deprecated_function()` does not support a custom message, which appears to be redundant here, as the method's DocBlock already indicates that it is unused as of WordPress 5.5.2.

Includes correcting the version in the `@deprecated` tag.

Follow-up to [49146], [49201].

Props dhruvik18, dhruvang21, rollybueno, SirLouen, SergeyBiryukov.
Fixes #63511.
Built from https://develop.svn.wordpress.org/trunk@60273


git-svn-id: https://core.svn.wordpress.org/trunk@59609 1a063a9b-81f0-0310-95a4-ce76da25c4cd