Build/Test Tools: Use `issues` instead of `issuesAndPullRequests`.
The `github.rest.search.issuesAndPullRequests()` method is deprecated and slated to be removed in September 2025.

The preferred way to search for pull requests is now to use `github.rest.search.issues` in combination with the `is:pr` or `type:pr` qualifiers.

Props abcd95.
Fixes #63278.
Built from https://develop.svn.wordpress.org/trunk@60314


git-svn-id: https://core.svn.wordpress.org/trunk@59650 1a063a9b-81f0-0310-95a4-ce76da25c4cd