Twenty Twenty-One: Fixes List Item Block Margin.
The margin was different depending on the front or back of the editor. This patch is a companion to the next commit incoming from issue 63549. That issue resolves the line-height inconsistency, and this one does not.
  
Props nidhidhandhukiya, ugyensupport, sandeepdahiya, rishabhwp, mleraygp, SirLouen, sabernhardt.  
Fixes #60196.

Built from https://develop.svn.wordpress.org/trunk@60366


git-svn-id: https://core.svn.wordpress.org/trunk@59702 1a063a9b-81f0-0310-95a4-ce76da25c4cd