Plugins: Ensure only plugin icons are floated.
There's no reason not to add a semantic class to the plugin icons on the update screen instead of making our CSS more complex, so let's do that and ensure that the existing `.plugin-icon` styles are scoped to `.plugin-card` on the install screen.

Partially reverts [60673].
Fixes #63120.


Built from https://develop.svn.wordpress.org/trunk@60680


git-svn-id: https://core.svn.wordpress.org/trunk@60016 1a063a9b-81f0-0310-95a4-ce76da25c4cd