Build/Test Tools: Pin Importer plugin version for PHP < 7.2.
The WordPress Importer plugin now requires PHP 7.2 or higher. The plugin’s unit tests were removed as of [59769] in favor of running only within the plugin’s repository on GitHub instead.

[59769] was made during the 6.8 release cycle, so all WP <= 6.7 branches still checkout the latest revision for the plugin before running the PHPUnit test suite.

This change modifies the `env:install` script to download version `0.9.0` of the plugin (which is the last version of the Importer plugin that supports PHP < 7.2) when a qualifying version is being tested.

Merges [60748] to the 5.3 branch.

Props @swissspidy.
See #63983.
Built from https://develop.svn.wordpress.org/branches/5.3@60761


git-svn-id: https://core.svn.wordpress.org/branches/5.3@60097 1a063a9b-81f0-0310-95a4-ce76da25c4cd