Upgrade/Install: WordPress text shows over logo on tab.
On the upgrade and repair screens, the logo link to wordpress.org was hidden using negative `text-indent`. This method makes the link impossible to click on using a mouse, and causes the text to overlay the image on keyboard focus.

On the install screen, if PHP is not running, the styles are not applied, leaving an unstyled link.

Remove the links for consistency with other screens following r47745. 

Props johnbillion, sabernhardt, joedolson.
Fixes #63950.
Built from https://develop.svn.wordpress.org/trunk@60898


git-svn-id: https://core.svn.wordpress.org/trunk@60234 1a063a9b-81f0-0310-95a4-ce76da25c4cd