Code Modernization: Use `str_contains()` in `get_avatar_data()`.
`str_contains()` was introduced in PHP 8.0 to perform a case-sensitive check indicating if the string to search in (haystack) contains the given substring (needle).

WordPress core includes a polyfill for `str_contains()` on PHP < 8.0 as of WordPress 5.9.

Follow-up to [52039], [55988], [56245], [60897].

Props TobiasBg.
See #63087.
Built from https://develop.svn.wordpress.org/trunk@60900


git-svn-id: https://core.svn.wordpress.org/trunk@60236 1a063a9b-81f0-0310-95a4-ce76da25c4cd