Upgrade/Install: Update sodium_compat to v1.23.0.
The previous version of sodium_compat was overly permissible with `sodium_base642bin()` when the `*_NO_PADDING` variants were specified, which was not compatible with `ext-sodium`. This has been fixed in version 1.22.0.

Version 1.23.0 includes some optimizations by replacing the array in the Curve25519 field element with 10 integer object properties instead. The result is a 7% to 12% speedup for the overall PHPUnit suite.

References:
* [https://github.com/paragonie/sodium_compat/releases/tag/v1.22.0 sodium_compat 1.22.0 release notes]
* [https://github.com/paragonie/sodium_compat/releases/tag/v1.23.0 sodium_compat 1.23.0 release notes]
* [https://github.com/paragonie/sodium_compat/compare/v1.21.2...v1.23.0 Full list of changes in sodium_compat 1.23.0]

Follow-up to [55699], [58752], [58753], [60787].

Props paragoninitiativeenterprises, SergeyBiryukov.
Fixes #64079.
Built from https://develop.svn.wordpress.org/trunk@60905


git-svn-id: https://core.svn.wordpress.org/trunk@60241 1a063a9b-81f0-0310-95a4-ce76da25c4cd