Media: A11y: Switch uploader toggle to `button` and set focus.
The control to switch between the browser uploader and the default uploader used a link with `target="_blank"`, but was driven by scripts. In a no-js context, this meant that the link could be used to open the browser uploader in a new tab. This is unnecessary, however, because the default uploader is not rendered when JS is not available.

On switching uploaders, browser focus was lost.

For more predictable keyboard and screen reader behavior, switch the media uploader toggle to a `button` element and set focus to the upload button in the new context.

Props dilipbheda, sabernhardt, rollybueno, westonruter, joedolson.
Fixes #63238.
Built from https://develop.svn.wordpress.org/trunk@61099


git-svn-id: https://core.svn.wordpress.org/trunk@60435 1a063a9b-81f0-0310-95a4-ce76da25c4cd