External Libraries: Upgrade PHPMailer to version 7.0.0.
This is exactly the same as version 6.11.1 but bumps the major version number to indicate the presence of a backward compatibility break in child classes.

Specifically, `lang()`, `setLanguage()`, and `$language` are now static, and should be called statically.

References:
* [https://github.com/PHPMailer/PHPMailer/releases/tag/v7.0.0 PHPMailer 7.0.0 release notes]
* [https://github.com/PHPMailer/PHPMailer/compare/v6.11.1...v7.0.0 Full list of changes in PHPMailer 7.0.0]

Follow-up to [54937], [55557], [56484], [57137], [59246], [59481], [60623], [60813], [60888].

Props SirLouen.
Fixes #64101.
Built from https://develop.svn.wordpress.org/trunk@61249


git-svn-id: https://core.svn.wordpress.org/trunk@60561 1a063a9b-81f0-0310-95a4-ce76da25c4cd