Database: Further correct MariaDB version check in `wpdb::has_cap()`.
On older PHP versions, MariaDB version is reported with the `5.5.5-` prefix, e.g. `5.5.5-10.6.24-MariaDB`.

This includes PHP 8.0.15 or earlier, as well as PHP 8.1.0—8.1.2, which was not accounted for previously.

This commit updates the condition for removing the `5.5.5-` prefix to include PHP 8.1.0—8.1.2.

Follow-up to [54384].

Props maximumsoftware, hbhalodia, SergeyBiryukov.
Fixes #64332.
Built from https://develop.svn.wordpress.org/trunk@61349


git-svn-id: https://core.svn.wordpress.org/trunk@60661 1a063a9b-81f0-0310-95a4-ce76da25c4cd